+++ app/config/storage.cpp
@@ -7,8 +7,7 @@
 
 namespace Config {
   Storage::Storage(const QString &filename) : changed(false) {
-    auto cfg_path = QStandardPaths::writableLocation(QStandardPaths::ConfigLocation);
-    auto default_filedir = QString("%1/mpz").arg(cfg_path);
+    auto default_filedir = configPath();
     auto default_filepath = QString("%1/%2").arg(default_filedir).arg(filename);
     qDebug() << "config file: " << default_filepath;
 
@@ -26,6 +25,11 @@ namespace Config {
 
   Storage::~Storage() {
     save();
+  }
+
+  QString Storage::configPath() {
+    auto cfg_path = QStandardPaths::writableLocation(QStandardPaths::ConfigLocation);
+    return QString("%1/mpz").arg(cfg_path);
   }
 
   Value Storage::get(const QString &key, bool *ok) const {